+++ tests/test_libs_mm_utils.py
@@ -1,10 +1,19 @@
 import unittest
 
-from cl_sii.libs.mm_utils import CustomMarshmallowDateField  # noqa: F401
+from cl_sii.libs.mm_utils import (  # noqa: F401
+    CustomMarshmallowDateField, validate_no_unexpected_input_fields,
+)
 
 
 class CustomMarshmallowDateFieldTest(unittest.TestCase):
 
     def test_x(self) -> None:
-        # TODO: implement!
+        # TODO: implement for 'CustomMarshmallowDateField'.
+        pass
+
+
+class FunctionsTest(unittest.TestCase):
+
+    def test_validate_no_unexpected_input_fields(self):
+        # TODO: implement for 'validate_no_unexpected_input_fields'.
         pass